+++ navis/graph/__init__.py
@@ -11,24 +11,63 @@
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #    GNU General Public License for more details.
 
-from .converters import (network2nx, network2igraph, neuron2igraph, nx2neuron,
-                         neuron2nx, neuron2KDTree, neuron2tangents)
-from .graph_utils import (classify_nodes, cut_skeleton, longest_neurite,
-                          split_into_fragments, reroot_skeleton, distal_to,
-                          dist_between, find_main_branchpoint,
-                          generate_list_of_childs, geodesic_matrix,
-                          node_label_sorting, _break_segments,
-                          _generate_segments, segment_length,
-                          _connected_components, rewire_skeleton,
-                          connected_subgraph, insert_nodes, remove_nodes,
-                          dist_to_root, skeleton_adjacency_matrix)
-from .clinic import (health_check)
+from .converters import (
+    network2nx,
+    network2igraph,
+    neuron2igraph,
+    nx2neuron,
+    neuron2nx,
+    neuron2KDTree,
+    neuron2tangents,
+    simplify_graph,
+)
+from .graph_utils import (
+    classify_nodes,
+    cut_skeleton,
+    longest_neurite,
+    split_into_fragments,
+    reroot_skeleton,
+    distal_to,
+    dist_between,
+    find_main_branchpoint,
+    generate_list_of_childs,
+    geodesic_matrix,
+    node_label_sorting,
+    _break_segments,
+    _generate_segments,
+    segment_length,
+    _connected_components,
+    rewire_skeleton,
+    connected_subgraph,
+    insert_nodes,
+    remove_nodes,
+    dist_to_root,
+    skeleton_adjacency_matrix,
+)
+from .clinic import health_check
 
 
-__all__ = ['cut_skeleton', 'longest_neurite', 'split_into_fragments',
-           'reroot_skeleton', 'distal_to', 'dist_between', 'segment_length',
-           'find_main_branchpoint', 'geodesic_matrix',
-           'rewire_skeleton', 'insert_nodes', 'remove_nodes', 'health_check',
-           'graph_utils', 'network2nx', 'network2igraph', 'neuron2igraph',
-           'nx2neuron', 'neuron2nx', 'neuron2KDTree', 'neuron2tangents',
-           'dist_to_root']
+__all__ = [
+    "cut_skeleton",
+    "longest_neurite",
+    "split_into_fragments",
+    "reroot_skeleton",
+    "distal_to",
+    "dist_between",
+    "segment_length",
+    "find_main_branchpoint",
+    "geodesic_matrix",
+    "rewire_skeleton",
+    "insert_nodes",
+    "remove_nodes",
+    "health_check",
+    "graph_utils",
+    "network2nx",
+    "network2igraph",
+    "neuron2igraph",
+    "nx2neuron",
+    "neuron2nx",
+    "neuron2KDTree",
+    "neuron2tangents",
+    "dist_to_root",
+]